fix(module): 向 componentDetection 声明组件层使用的 @nuxt/ui 组件
movk 组件层内部渲染的 @nuxt/ui 组件(如 MDataTable 内的 UTable)不会被
@nuxt/ui 的 componentDetection 扫描到(其忽略 node_modules),导致这些组件的
主题类(如 Table 的 min-w-full)缺失 @source 而不生成 CSS。

将组件清单抽到 src/utils/ui-components.ts,module 与 unplugin 共享:module 据此
(去 U 前缀)在消费方启用检测(true 或数组)时并入 componentDetection。抽离避免
module 直接从 unplugin 引入而把路径相关的 runtimeDir 卷入 dist/shared 分块,
否则会破坏 Vue 插件对 runtime composables 的解析。unplugin 保留 UI_COMPONENTS
再导出以维持公开 API。
