+++ src/utils/ui-components.ts
@@ -1,0 +1,9 @@
+/** movk 组件层从 `#components` 具名导入的 @nuxt/ui 组件（去 U 前缀映射到 @nuxt/ui/components/<Name>.vue） */
+export const UI_COMPONENTS = [
+  'UAccordion',
+  'UButton', 'UBadge', 'UCheckbox', 'UCollapsible', 'UForm', 'UDropdownMenu', 'UIcon', 'UInput',
+  'UModal', 'UPagination', 'USelect', 'UPopover', 'UCalendar', 'UColorPicker', 'UTable', 'UFormField',
+  'UTooltip', 'UInputNumber', 'USwitch', 'UTextarea', 'USlider', 'UPinInput', 'UInputTags',
+  'UFileUpload', 'USelectMenu', 'UInputMenu', 'UCheckboxGroup', 'URadioGroup', 'UInputDate',
+  'UInputTime', 'UListbox'
+] as const
